skillbridge_backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.10.0
attrs==25.4.0
Authlib==1.6.5
bandit==1.7.6
billiard==4.2.2
black==23.11.0
build==1.3.0
celery==5.3.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.3.2
cryptography==46.0.3
Deprecated==1.3.1
distro==1.9.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dparse==0.6.4
drf-spectacular==0.29.0
factory-boy==3.3.0
Faker==37.12.0
filelock==3.20.0
flake8==6.1.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
isort==5.12.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.7.1
mypy_extensions==1.1.0
nltk==3.9.2
openai==1.3.0
packaging==25.0
pathspec==0.12.1
pip-tools==7.5.1
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.1.3
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyflakes==3.1.0
PyGithub==2.1.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.0
pyproject_hooks==1.2.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
regex==2025.11.3
requests==2.31.0
rich==14.2.0
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safety==3.6.2
safety-schemas==0.0.16
sentry-sdk==1.38.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
stevedore==5.5.0
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
wrapt==2.0.0
